Backend/ReSift API/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.8.0
MarkupSafe==2.1.0
newspaper3k==0.2.8
nltk==3.7
Pillow==9.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tinysegmenter==0.3
tldextract==3.2.0
tqdm==4.63.0
urllib3==1.26.8
Werkzeug==2.0.3
favicon==0.7.0
